refactor: UI component 분리
