+++ src/basic/components/cart/index.js
@@ -1,0 +1,5 @@
+/**
+ * Cart 컴포넌트들의 barrel export
+ */
+export { createCartDisplay } from './CartDisplay.js';
+export { renderCartItem } from './CartItem.js';
