+++ src/basic/components/index.js
@@ -1,0 +1,18 @@
+/**
+ * 모든 컴포넌트들의 barrel export
+ */
+
+// Layout 컴포넌트들
+export { createHeader, createGridContainer } from './layout/index.js';
+
+// Product 컴포넌트들
+export { createProductSelector, createStockInfo } from './product/index.js';
+
+// Cart 컴포넌트들
+export { createCartDisplay, renderCartItem } from './cart/index.js';
+
+// Order 컴포넌트들
+export { createRightColumn, renderDiscountInfo, renderLoyaltyPoints } from './order/index.js';
+
+// Modal 컴포넌트들
+export { createManualOverlay, createManualToggle, createManualColumn } from './modal/index.js';
