+++ src/basic/components/layout/index.js
@@ -1,0 +1,5 @@
+/**
+ * Layout 컴포넌트들의 barrel export
+ */
+export { createHeader } from './Header.js';
+export { createGridContainer } from './GridContainer.js';
