+++ src/basic/components/modal/index.js
@@ -1,0 +1,6 @@
+/**
+ * Modal 컴포넌트들의 barrel export
+ */
+export { createManualOverlay } from './ManualOverlay.js';
+export { createManualToggle } from './ManualToggle.js';
+export { createManualColumn } from './ManualColumn.js';
