+++ src/basic/components/order/index.js
@@ -1,0 +1,6 @@
+/**
+ * Order 컴포넌트들의 barrel export
+ */
+export { createRightColumn } from './OrderSummary.js';
+export { renderDiscountInfo } from './DiscountInfo.js';
+export { renderLoyaltyPoints } from './LoyaltyPoints.js';
