+++ src/basic/components/product/index.js
@@ -1,0 +1,5 @@
+/**
+ * Product 컴포넌트들의 barrel export
+ */
+export { createProductSelector } from './ProductSelector.js';
+export { createStockInfo } from './StockInfo.js';
