+++ src/basic/main.basic.js
@@ -16,17 +16,47 @@ import {
 } from './services/product/ProductService.js';
 
 // DiscountService import
-import {
-  calculateTotalDiscountRate,
-  createDiscountInfo,
-  calculateSavedAmount,
-} from './services/discount/DiscountService.js';
+import { calculateTotalDiscountRate } from './services/discount/DiscountService.js';
 
 // PointService import
 import { createPointInfo } from './services/point/PointService.js';
 
 // CartService import
 import { createInitialCartState, addItemToCart } from './services/cart/CartService.js';
+
+// Components import (새로운 폴더 구조)
+import {
+  createHeader,
+  createGridContainer,
+  createProductSelector,
+  createStockInfo,
+  createCartDisplay,
+  createRightColumn,
+  createManualOverlay,
+  createManualToggle,
+  createManualColumn,
+  renderCartItem,
+  renderDiscountInfo,
+  renderLoyaltyPoints,
+} from './components/index.js';
+
+// Utils import
+import {
+  createAddToCartButton,
+  createSelectorContainer,
+  createLeftColumn,
+} from './utils/UIRenderer.js';
+
+// Renderers import
+import {
+  renderProductOptions,
+  renderOrderSummaryDetails,
+  renderTuesdaySpecial,
+  renderTotalAmount,
+  renderItemCount,
+  renderStockMessages,
+  updateCartPrices,
+} from './utils/renderers/index.js';
 
 // 전역 변수들 (명명 규칙 적용)
 let productList;
@@ -73,176 +103,38 @@ function main() {
 
   const root = document.getElementById('app');
 
-  const header = document.createElement('div');
-  header.className = 'mb-8';
-  header.innerHTML = `
-    <h1 class="text-xs font-medium tracking-extra-wide uppercase mb-2">🛒 Hanghae Online Store</h1>
-    <div class="text-5xl tracking-tight leading-none">Shopping Cart</div>
-    <p id="item-count" class="text-sm text-gray-500 font-normal mt-3">🛍️ 0 items in cart</p>
-  `;
+  // UI 컴포넌트들 생성
+  const header = createHeader();
 
-  productSelector = document.createElement('select');
-  productSelector.id = 'product-select';
-  productSelector.className = 'w-full p-3 border border-gray-300 rounded-lg text-base mb-3';
+  productSelector = createProductSelector();
+  addToCartButton = createAddToCartButton();
+  stockInfoElement = createStockInfo();
 
-  const leftColumn = document.createElement('div');
-  leftColumn['className'] = 'bg-white border border-gray-200 p-8 overflow-y-auto';
+  const selectorContainer = createSelectorContainer(
+    productSelector,
+    addToCartButton,
+    stockInfoElement,
+  );
+  cartDisplayElement = createCartDisplay();
 
-  const gridContainer = document.createElement('div');
-  gridContainer.className =
-    'grid grid-cols-1 lg:grid-cols-[1fr_360px] gap-6 flex-1 overflow-hidden';
+  const leftColumn = createLeftColumn(selectorContainer, cartDisplayElement);
+  const rightColumn = createRightColumn();
 
-  stockInfoElement = document.createElement('div');
-  stockInfoElement.id = 'stock-status';
-  stockInfoElement.className = 'text-xs text-red-500 mt-3 whitespace-pre-line';
+  const manualOverlay = createManualOverlay();
+  const manualToggle = createManualToggle();
+  const manualColumn = createManualColumn();
 
-  addToCartButton = document.createElement('button');
-  addToCartButton.id = 'add-to-cart';
-  addToCartButton.innerHTML = 'Add to Cart';
-  addToCartButton.className =
-    'w-full py-3 bg-black text-white text-sm font-medium uppercase tracking-wider hover:bg-gray-800 transition-all';
+  const gridContainer = createGridContainer(leftColumn, rightColumn);
 
-  // 상품 선택/추가/재고 표시 컨테이너
-  const selectorContainer = document.createElement('div');
-  selectorContainer.className = 'mb-6 pb-6 border-b border-gray-200';
-  selectorContainer.appendChild(productSelector);
-  selectorContainer.appendChild(addToCartButton);
-  selectorContainer.appendChild(stockInfoElement);
-
-  cartDisplayElement = document.createElement('div');
-  cartDisplayElement.id = 'cart-items';
-
-  leftColumn.appendChild(selectorContainer);
-  leftColumn.appendChild(cartDisplayElement);
-
-  // 오른쪽 컬럼(주문 요약) 생성
-  const rightColumn = document.createElement('div');
-  rightColumn.className = 'bg-black text-white p-8 flex flex-col';
-  rightColumn.innerHTML = `
-    <h2 class="text-xs font-medium mb-5 tracking-extra-wide uppercase">Order Summary</h2>
-    <div class="flex-1 flex flex-col">
-      <div id="summary-details" class="space-y-3"></div>
-      <div class="mt-auto">
-        <div id="discount-info" class="mb-4"></div>
-        <div id="cart-total" class="pt-5 border-t border-white/10">
-          <div class="flex justify-between items-baseline">
-            <span class="text-sm uppercase tracking-wider">Total</span>
-            <div class="text-2xl tracking-tight">₩0</div>
-          </div>
-          <div id="loyalty-points" class="text-xs text-blue-400 mt-2 text-right">적립 포인트: 0p</div>
-        </div>
-        <div id="tuesday-special" class="mt-4 p-3 bg-white/10 rounded-lg hidden">
-          <div class="flex items-center gap-2">
-            <span class="text-2xs">🎉</span>
-            <span class="text-xs uppercase tracking-wide">Tuesday Special 10% Applied</span>
-          </div>
-        </div>
-      </div>
-    </div>
-    <button class="w-full py-4 bg-white text-black text-sm font-normal uppercase tracking-super-wide cursor-pointer mt-6 transition-all hover:-translate-y-0.5 hover:shadow-lg hover:shadow-black/30">
-      Proceed to Checkout
-    </button>
-    <p class="mt-4 text-2xs text-white/60 text-center leading-relaxed">
-      Free shipping on all orders.<br>
-      <span id="points-notice">Earn loyalty points with purchase.</span>
-    </p>
-  `;
-  orderSummaryElement = rightColumn.querySelector('#cart-total');
-
-  // 이용 안내(오버레이) 관련 요소 생성
-  const manualOverlay = document.createElement('div');
-  manualOverlay.className = 'fixed inset-0 bg-black/50 z-40 hidden transition-opacity duration-300';
-  manualOverlay.onclick = function (e) {
-    if (e.target === manualOverlay) {
-      manualOverlay.classList.add('hidden');
-      manualColumn.classList.add('translate-x-full');
-    }
-  };
-
-  const manualToggle = document.createElement('button');
-  manualToggle.onclick = function () {
-    manualOverlay.classList.toggle('hidden');
-    manualColumn.classList.toggle('translate-x-full');
-  };
-  manualToggle.className =
-    'fixed top-4 right-4 bg-black text-white p-3 rounded-full hover:bg-gray-900 transition-colors z-50';
-  manualToggle.innerHTML = `
-    <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-      <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M13 16h-1v-4h-1m1-4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"></path>
-    </svg>
-  `;
-
-  const manualColumn = document.createElement('div');
-  manualColumn.className =
-    'fixed right-0 top-0 h-full w-80 bg-white shadow-2xl p-6 overflow-y-auto z-50 transform translate-x-full transition-transform duration-300';
-  manualColumn.innerHTML = `
-    <button class="absolute top-4 right-4 text-gray-500 hover:text-black" onclick="document.querySelector('.fixed.inset-0').classList.add('hidden'); this.parentElement.classList.add('translate-x-full')">
-      <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"></path>
-      </svg>
-    </button>
-    <h2 class="text-xl font-bold mb-4">📖 이용 안내</h2>
-    <div class="mb-6">
-      <h3 class="text-base font-bold mb-3">💰 할인 정책</h3>
-      <div class="space-y-3">
-        <div class="bg-gray-100 rounded-lg p-3">
-          <p class="font-semibold text-sm mb-1">개별 상품</p>
-          <p class="text-gray-700 text-xs pl-2">
-            • 키보드 10개↑: 10%<br>
-            • 마우스 10개↑: 15%<br>
-            • 모니터암 10개↑: 20%<br>
-            • 스피커 10개↑: 25%
-          </p>
-        </div>
-        <div class="bg-gray-100 rounded-lg p-3">
-          <p class="font-semibold text-sm mb-1">전체 수량</p>
-          <p class="text-gray-700 text-xs pl-2">• 30개 이상: 25%</p>
-        </div>
-        <div class="bg-gray-100 rounded-lg p-3">
-          <p class="font-semibold text-sm mb-1">특별 할인</p>
-          <p class="text-gray-700 text-xs pl-2">
-            • 화요일: +10%<br>
-            • ⚡번개세일: 20%<br>
-            • 💝추천할인: 5%
-          </p>
-        </div>
-      </div>
-    </div>
-    <div class="mb-6">
-      <h3 class="text-base font-bold mb-3">🎁 포인트 적립</h3>
-      <div class="space-y-3">
-        <div class="bg-gray-100 rounded-lg p-3">
-          <p class="font-semibold text-sm mb-1">기본</p>
-          <p class="text-gray-700 text-xs pl-2">• 구매액의 0.1%</p>
-        </div>
-        <div class="bg-gray-100 rounded-lg p-3">
-          <p class="font-semibold text-sm mb-1">추가</p>
-          <p class="text-gray-700 text-xs pl-2">
-            • 화요일: 2배<br>
-            • 키보드+마우스: +50p<br>
-            • 풀세트: +100p<br>
-            • 10개↑: +20p / 20개↑: +50p / 30개↑: +100p
-          </p>
-        </div>
-      </div>
-    </div>
-    <div class="border-t border-gray-200 pt-4 mt-4">
-      <p class="text-xs font-bold mb-1">💡 TIP</p>
-      <p class="text-2xs text-gray-600 leading-relaxed">
-        • 화요일 대량구매 = MAX 혜택<br>
-        • ⚡+💝 중복 가능<br>
-        • 상품4 = 품절
-      </p>
-    </div>
-  `;
-
-  gridContainer.appendChild(leftColumn);
-  gridContainer.appendChild(rightColumn);
+  // DOM에 요소들 추가
   manualOverlay.appendChild(manualColumn);
   root.appendChild(header);
   root.appendChild(gridContainer);
   root.appendChild(manualToggle);
   root.appendChild(manualOverlay);
+
+  // orderSummaryElement 참조 설정
+  orderSummaryElement = rightColumn.querySelector('#cart-total');
 
   // 상품 옵션, 장바구니, 가격 등 초기 렌더링
   updateProductOptions();
@@ -354,106 +246,15 @@ function calculateTotalDiscount(subTot, itemCount, currentAmount) {
 
 // 주문 요약 상세 내역 갱신
 function updateOrderSummary(cartItems, subTot, itemCount, itemDiscounts) {
-  const summaryDetails = document.getElementById('summary-details');
-  summaryDetails.innerHTML = '';
-
-  if (subTot > 0) {
-    // 각 상품별 정보 표시
-    for (let i = 0; i < cartItems.length; i++) {
-      let curItem;
-      for (let j = 0; j < productList.length; j++) {
-        if (productList[j].id === cartItems[i].id) {
-          curItem = productList[j];
-          break;
-        }
-      }
-
-      const quantityElem = cartItems[i].querySelector('.quantity-number');
-      const quantity = parseInt(quantityElem.textContent);
-      const itemTotal = curItem.price * quantity;
-
-      summaryDetails.innerHTML += `
-        <div class="flex justify-between text-xs tracking-wide text-gray-400">
-          <span>${curItem.name} x ${quantity}</span>
-          <span>₩${itemTotal.toLocaleString()}</span>
-        </div>
-      `;
-    }
-
-    // 소계 표시
-    summaryDetails.innerHTML += `
-      <div class="border-t border-white/10 my-3"></div>
-      <div class="flex justify-between text-sm tracking-wide">
-        <span>Subtotal</span>
-        <span>₩${subTot.toLocaleString()}</span>
-      </div>
-    `;
-
-    // 할인 정보 표시 - DiscountService 사용
-    const discountInfo = createDiscountInfo(itemDiscounts, itemCount);
-    discountInfo.forEach(function (discount) {
-      const colorClass = discount.type === 'tuesday' ? 'text-purple-400' : 'text-green-400';
-      const icon = discount.type === 'tuesday' ? '🌟' : discount.type === 'bulk' ? '🎉' : '';
-      summaryDetails.innerHTML += `
-        <div class="flex justify-between text-sm tracking-wide ${colorClass}">
-          <span class="text-xs">${icon} ${discount.name}</span>
-          <span class="text-xs">-${discount.rate}%</span>
-        </div>
-      `;
-    });
-
-    // 배송비 표시
-    summaryDetails.innerHTML += `
-      <div class="flex justify-between text-sm tracking-wide text-gray-400">
-        <span>Shipping</span>
-        <span>Free</span>
-      </div>
-    `;
-  }
+  renderOrderSummaryDetails(cartItems, productList, subTot, itemDiscounts);
 }
 
 // 상품 선택 옵션 렌더링 및 재고 상태 표시
 function updateProductOptions() {
-  let opt;
-  let discountText;
-
-  productSelector.innerHTML = '';
+  renderProductOptions(productSelector, productList);
 
   // ProductService의 getTotalStock 함수 사용
   const totalStock = getTotalStock(productList);
-
-  // 각 상품별 옵션 생성
-  for (let i = 0; i < productList.length; i++) {
-    (function () {
-      const item = productList[i];
-      opt = document.createElement('option');
-      opt.value = item.id;
-      discountText = '';
-
-      if (item.onSale) discountText += ' ⚡SALE';
-      if (item.suggestSale) discountText += ' 💝추천';
-
-      if (item.quantity === 0) {
-        opt.textContent = `${item.name} - ${item.price}원 (품절)${discountText}`;
-        opt.disabled = true;
-        opt.className = 'text-gray-400';
-      } else {
-        if (item.onSale && item.suggestSale) {
-          opt.textContent = `⚡💝${item.name} - ${item.originalPrice}원 → ${item.price}원 (25% SUPER SALE!)`;
-          opt.className = 'text-purple-600 font-bold';
-        } else if (item.onSale) {
-          opt.textContent = `⚡${item.name} - ${item.originalPrice}원 → ${item.price}원 (20% SALE!)`;
-          opt.className = 'text-red-500 font-bold';
-        } else if (item.suggestSale) {
-          opt.textContent = `💝${item.name} - ${item.originalPrice}원 → ${item.price}원 (5% 추천할인!)`;
-          opt.className = 'text-blue-500 font-bold';
-        } else {
-          opt.textContent = `${item.name} - ${item.price}원${discountText}`;
-        }
-      }
-      productSelector.appendChild(opt);
-    })();
-  }
 
   if (totalStock < UI_CONSTANTS.TOTAL_STOCK_THRESHOLD) {
     productSelector.style.borderColor = 'orange';
@@ -464,7 +265,6 @@ function updateProductOptions() {
 
 // 장바구니, 할인, 포인트 등 계산 및 화면 갱신
 function calculateCartSummary() {
-  let savedAmount;
   let points;
   let previousCount;
 
@@ -493,22 +293,17 @@ function calculateCartSummary() {
   const discRate = discountRate;
 
   // 화요일 특별 할인 UI 표시
-  const tuesdaySpecial = document.getElementById('tuesday-special');
-  if (isTuesday && totalAmount > 0) {
-    tuesdaySpecial.classList.remove('hidden');
-  } else {
-    tuesdaySpecial.classList.add('hidden');
-  }
+  renderTuesdaySpecial(isTuesday, totalAmount);
+
   // 장바구니 수량 표시 갱신
-  document.getElementById('item-count').textContent = `🛍️ ${itemCount} items in cart`;
+  renderItemCount(itemCount);
 
   // 주문 요약(상품별, 할인, 배송 등) 갱신
   updateOrderSummary(cartItems, subTot, itemCount, itemDiscounts);
+
   // 총 결제 금액 표시 갱신
-  const totalDiv = orderSummaryElement.querySelector('.text-2xl');
-  if (totalDiv) {
-    totalDiv.textContent = `₩${Math.round(totalAmount).toLocaleString()}`;
-  }
+  renderTotalAmount(totalAmount, orderSummaryElement);
+
   // 적립 포인트 표시 갱신 - PointService 사용
   const loyaltyPointsDiv = document.getElementById('loyalty-points');
   if (loyaltyPointsDiv) {
@@ -527,41 +322,21 @@ function calculateCartSummary() {
     const pointInfo = createPointInfo(totalAmount, cartItems);
     points = pointInfo.totalPoints;
 
-    if (points > 0) {
-      loyaltyPointsDiv.textContent = `적립 포인트: ${points}p`;
-      loyaltyPointsDiv.style.display = 'block';
-    } else {
-      loyaltyPointsDiv.textContent = '적립 포인트: 0p';
-      loyaltyPointsDiv.style.display = 'block';
-    }
+    renderLoyaltyPoints(points, pointInfo);
   }
-  // 할인 정보 표시 갱신
-  const discountInfoDiv = document.getElementById('discount-info');
-  discountInfoDiv.innerHTML = '';
 
-  if (discRate > 0 && totalAmount > 0) {
-    savedAmount = calculateSavedAmount(originalTotal, totalAmount);
-    discountInfoDiv.innerHTML = `
-      <div class="bg-green-500/20 rounded-lg p-3">
-        <div class="flex justify-between items-center mb-1">
-          <span class="text-xs uppercase tracking-wide text-green-400">총 할인율</span>
-          <span class="text-sm font-medium text-green-400">${(discRate * 100).toFixed(1)}%</span>
-        </div>
-        <div class="text-2xs text-gray-300">₩${Math.round(
-          savedAmount,
-        ).toLocaleString()} 할인되었습니다</div>
-      </div>
-    `;
-  }
+  // 할인 정보 표시 갱신
+  renderDiscountInfo(discRate, originalTotal, totalAmount);
+
   // 장바구니 수량 변화 애니메이션 표시
   const itemCountElement = document.getElementById('item-count');
   if (itemCountElement) {
     previousCount = parseInt(itemCountElement.textContent.match(/\d+/) || 0);
-    itemCountElement.textContent = `🛍️ ${itemCount} items in cart`;
     if (previousCount !== itemCount) {
       itemCountElement.setAttribute('data-changed', 'true');
     }
   }
+
   // 재고 부족/품절 안내 메시지 갱신
   updateStockMessages();
 
@@ -612,60 +387,11 @@ function updateStockMessages() {
   // 품절 상품 조회
   const outOfStockProducts = getOutOfStockProducts(productList);
 
-  let stockMsg = '';
-
-  // 재고 부족 상품 메시지
-  lowStockProducts.forEach((item) => {
-    stockMsg += `${item.name}: 재고 부족 (${item.quantity}개 남음)\n`;
-  });
-
-  // 품절 상품 메시지
-  outOfStockProducts.forEach((item) => {
-    stockMsg += `${item.name}: 품절\n`;
-  });
-
-  stockInfoElement.textContent = stockMsg;
-}
-
-// 상품 가격 렌더링 로직 분리
-function renderProductPrice(product, priceDiv, nameDiv) {
-  if (product.onSale && product.suggestSale) {
-    priceDiv.innerHTML = `<span class="line-through text-gray-400">₩${product.originalPrice.toLocaleString()}</span> <span class="text-purple-600">₩${product.price.toLocaleString()}</span>`;
-    nameDiv.textContent = `⚡💝${product.name}`;
-  } else if (product.onSale) {
-    priceDiv.innerHTML = `<span class="line-through text-gray-400">₩${product.originalPrice.toLocaleString()}</span> <span class="text-red-500">₩${product.price.toLocaleString()}</span>`;
-    nameDiv.textContent = `⚡${product.name}`;
-  } else if (product.suggestSale) {
-    priceDiv.innerHTML = `<span class="line-through text-gray-400">₩${product.originalPrice.toLocaleString()}</span> <span class="text-blue-500">₩${product.price.toLocaleString()}</span>`;
-    nameDiv.textContent = `💝${product.name}`;
-  } else {
-    priceDiv.textContent = `₩${product.price.toLocaleString()}`;
-    nameDiv.textContent = product.name;
-  }
+  renderStockMessages(lowStockProducts, outOfStockProducts, stockInfoElement);
 }
 
 // 장바구니 내 상품 가격/이름 갱신 및 전체 금액 재계산
-function updateCartPrices() {
-  const cartItems = cartDisplayElement.children;
-
-  const productMap = productList.reduce((map, product) => {
-    map[product.id] = product;
-    return map;
-  }, {});
-
-  // 각 카트 아이템을 순회하며 가격/이름 업데이트 (productMap 사용)
-  for (const cartItem of cartItems) {
-    const product = productMap[cartItem.id];
-
-    if (product) {
-      const priceDiv = cartItem.querySelector('.text-lg');
-      const nameDiv = cartItem.querySelector('h3');
-      renderProductPrice(product, priceDiv, nameDiv);
-    }
-  }
-
-  calculateCartSummary();
-}
+// updateCartPrices 함수는 utils/renderers/CartRenderer.js에서 import됨
 
 // CartService를 사용한 장바구니 아이템 추가 함수
 function addItemToCartUI(productId, quantity = 1) {
@@ -690,65 +416,7 @@ function renderCartItems() {
     const product = productService.getProductById(item.id);
     if (!product) return;
 
-    const newItem = document.createElement('div');
-    newItem.id = item.id;
-    newItem.className =
-      'grid grid-cols-[80px_1fr_auto] gap-5 py-5 border-b border-gray-100 first:pt-0 last:border-b-0 last:pb-0';
-    newItem.innerHTML = `
-      <div class="w-20 h-20 bg-gradient-black relative overflow-hidden">
-        <div class="absolute top-1/2 left-1/2 w-[60%] h-[60%] bg-white/10 -translate-x-1/2 -translate-y-1/2 rotate-45"></div>
-      </div>
-      <div>
-        <h3 class="text-base font-normal mb-1 tracking-tight">${
-          product.onSale && product.suggestSale
-            ? '⚡💝'
-            : product.onSale
-            ? '⚡'
-            : product.suggestSale
-            ? '💝'
-            : ''
-        }${product.name}</h3>
-        <p class="text-xs text-gray-500 mb-0.5 tracking-wide">PRODUCT</p>
-        <p class="text-xs text-black mb-3">${
-          product.onSale || product.suggestSale
-            ? `<span class="line-through text-gray-400">₩${product.originalPrice.toLocaleString()}</span> <span class="${
-                product.onSale && product.suggestSale
-                  ? 'text-purple-600'
-                  : product.onSale
-                  ? 'text-red-500'
-                  : 'text-blue-500'
-              }">₩${product.price.toLocaleString()}</span>`
-            : `₩${product.price.toLocaleString()}`
-        }</p>
-        <div class="flex items-center gap-4">
-          <button class="quantity-change w-6 h-6 border border-black bg-white text-sm flex items-center justify-center transition-all hover:bg-black hover:text-white" data-product-id="${
-            item.id
-          }" data-change="-1">−</button>
-          <span class="quantity-number text-sm font-normal min-w-[20px] text-center tabular-nums">${
-            item.quantity
-          }</span>
-          <button class="quantity-change w-6 h-6 border border-black bg-white text-sm flex items-center justify-center transition-all hover:bg-black hover:text-white" data-product-id="${
-            item.id
-          }" data-change="1">+</button>
-        </div>
-      </div>
-      <div class="text-right">
-        <div class="text-lg mb-2 tracking-tight tabular-nums">${
-          product.onSale || product.suggestSale
-            ? `<span class="line-through text-gray-400">₩${product.originalPrice.toLocaleString()}</span> <span class="${
-                product.onSale && product.suggestSale
-                  ? 'text-purple-600'
-                  : product.onSale
-                  ? 'text-red-500'
-                  : 'text-blue-500'
-              }">₩${product.price.toLocaleString()}</span>`
-            : `₩${product.price.toLocaleString()}`
-        }</div>
-        <a class="remove-item text-2xs text-gray-500 uppercase tracking-wider cursor-pointer transition-colors border-b border-transparent hover:text-black hover:border-black" data-product-id="${
-          item.id
-        }">Remove</a>
-      </div>
-    `;
+    const newItem = renderCartItem(item, product);
     cartDisplayElement.appendChild(newItem);
   });
 }
